Automatically backed up by Learn

+++ db/migrate/001_create_users_table.rb
@@ -3,7 +3,7 @@ class CreateUserTable < ActiveRecord::Migration
     create_table users do |t|
       t.string :name
       t.string :password
-      t.integer :balance
+      t.decimal :balance
     end
   end
 end
